#11 Update

+++ SinFronteras/src/main/java/co/edu/sinfronteras/dao/UserDao.java
@@ -7,5 +7,5 @@ import co.edu.sinfronteras.model.User;
 public interface UserDao {
    void save(User user);
    List<User> list();
-   void eliminar(Long userId);
+   void eliminar(Integer userId);
 }
